Fix Balance behaviour

+++ src/services/database/migrations/20210228062006-create-service.js
@@ -19,7 +19,7 @@ module.exports = {
                 ]
             },
             cost: {
-                type: DataTypes.FLOAT,
+                type: DataTypes.DECIMAL(20, 2),
                 defaultValue: 0.0
             },
             status: {
